client: move landing page content to new LandingScreen component
